Move base shaders into one module
